Move implementation of processor config utils to cpp file
