+++ libminifi/include/utils/ProcessorConfigUtils.h
@@ -21,7 +21,6 @@
 #include <string>
 #include <set>
 
-#include "utils/StringUtils.h"
 #include "core/ProcessContext.h"
 
 namespace org {
@@ -30,60 +29,13 @@ namespace nifi {
 namespace minifi {
 namespace utils {
 
-std::string getRequiredPropertyOrThrow(const core::ProcessContext* context, const std::string& property_name) {
-  std::string value;
-  if (!context->getProperty(property_name, value)) {
-    throw std::runtime_error(property_name + " property missing or invalid");
-  }
-  return value;
-}
-
-std::vector<std::string> listFromCommaSeparatedProperty(const core::ProcessContext* context, const std::string& property_name) {
-  std::string property_string;
-  context->getProperty(property_name, property_string);
-  return utils::StringUtils::splitAndTrim(property_string, ",");
-}
-
-std::vector<std::string> listFromRequiredCommaSeparatedProperty(const core::ProcessContext* context, const std::string& property_name) {
-  return utils::StringUtils::splitAndTrim(getRequiredPropertyOrThrow(context, property_name), ",");
-}
-
-bool parseBooleanPropertyOrThrow(core::ProcessContext* context, const std::string& property_name) {
-  const std::string value_str = getRequiredPropertyOrThrow(context, property_name);
-  utils::optional<bool> maybe_value = utils::StringUtils::toBool(value_str);
-  if (!maybe_value) {
-    throw std::runtime_error(property_name + " property is invalid: value is " + value_str);
-  }
-  return maybe_value.value();
-}
-
-std::chrono::milliseconds parseTimePropertyMSOrThrow(core::ProcessContext* context, const std::string& property_name) {
-  core::TimeUnit unit;
-  uint64_t time_value_ms;
-  const std::string value_str = getRequiredPropertyOrThrow(context, property_name);
-  if (!core::Property::StringToTime(value_str, time_value_ms, unit) || !core::Property::ConvertTimeUnitToMS(time_value_ms, unit, time_value_ms)) {
-    throw std::runtime_error(property_name + " property is invalid: value is " + value_str);
-  }
-  return std::chrono::milliseconds(time_value_ms);
-}
-
-utils::optional<uint64_t> getOptionalUintProperty(const core::ProcessContext& context, const std::string& property_name) {
-  uint64_t value;
-  if (context.getProperty(property_name, value)) {
-    return { value };
-  }
-  return utils::nullopt;
-}
-
-std::string parsePropertyWithAllowableValuesOrThrow(const core::ProcessContext& context, const std::string& property_name, const std::set<std::string> allowable_values) {
-  std::string value;
-  if (!context.getProperty(property_name, value)
-      || value.empty()
-      || allowable_values.find(value) == allowable_values.end()) {
-    throw Exception(PROCESS_SCHEDULE_EXCEPTION, property_name + " property missing or invalid");
-  }
-  return value;
-}
+std::string getRequiredPropertyOrThrow(const core::ProcessContext* context, const std::string& property_name);
+std::vector<std::string> listFromCommaSeparatedProperty(const core::ProcessContext* context, const std::string& property_name);
+std::vector<std::string> listFromRequiredCommaSeparatedProperty(const core::ProcessContext* context, const std::string& property_name);
+bool parseBooleanPropertyOrThrow(core::ProcessContext* context, const std::string& property_name);
+std::chrono::milliseconds parseTimePropertyMSOrThrow(core::ProcessContext* context, const std::string& property_name);
+utils::optional<uint64_t> getOptionalUintProperty(const core::ProcessContext& context, const std::string& property_name);
+std::string parsePropertyWithAllowableValuesOrThrow(const core::ProcessContext& context, const std::string& property_name, const std::set<std::string>& allowable_values);
 
 }  // namespace utils
 }  // namespace minifi
